requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0

# Image processing and generation
diffusers>=0.21.0
Pillow>=9.0.0

# Data manipulation
pandas>=1.5.0
numpy>=1.21.0

# Configuration and logging
hydra-core>=1.3.0
wandb>=0.15.0
omegaconf>=2.3.0

# Additional utilities
einops>=0.6.0
tqdm>=4.64.0

# OCR dependencies (if using OCR features)
pytesseract>=0.3.10
easyocr>=1.7.0

# String matching for OCR
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# CLIP models
open-clip-torch>=2.20.0

# Image augmentation
albumentations>=1.3.0

# Development tools (optional)
jupyter>=1.0.0
ipykernel>=6.0.0